Improve intro script
- Make visually easier to read.
- Rename to z_intro_script.py
